164797996-feature(comment): fix delete modal and responsiveness
