+++ src/components/presentations/UserArticles/author-articles.scss
@@ -3,7 +3,7 @@ h3, p {
     font-family: $font-stack;
 }
 .article-container {
-    overflow: auto;
+    margin-top: 10%;
 }
 .profile-article-list {
     text-align: center;
