+++ src/components/shared/NavBar/navbar.scss
@@ -140,7 +140,7 @@
   }
 }
 
- @media screen and (max-width: 700px) {
+ @media screen and (max-width: 768px) {
    .nav-header {
     .ui.menu {
       width: 100%;
